requirements.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.7.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.2
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==3.0
django-phonenumber-field==4.0.0
djangorestframework==3.10.3
geographiclib==1.50
geopy==1.20.0
googletrans==2.4.0
gTTS==2.0.4
gTTS-token==1.1.3
http-ece==1.1.0
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
phonenumbers==8.11.1
psycopg2-binary==2.8.4
py-vapid==1.7.0
pycparser==2.19
pytz==2019.3
pywebpush==1.10.1
redis==3.4.1
regex==2020.1.8
requests==2.22.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
splinter==0.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==2.1.0
